w3c-reports/12092014/mac-wk2/runned-tests/2dcontext/drawing-images-to-the-canvas/2d.drawImage.negativesource-expected.txt

2d.drawImage.negativesource
Negative source width/height represents the correct rectangle
Actual output:
Expected output:


FAIL Negative source width/height represents the correct rectangle assert_approx_equals: Red channel of the pixel at (1, 1) expected 0 +/- 2 but got 255(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_approx_equals@http://localhost:8800/resources/testharness.js:425:15
_assertPixelApprox@http://localhost:8800/common/canvas-tests.js:47:25
http://localhost:8800/2dcontext/drawing-images-to-the-canvas/2d.drawImage.negativesource.html:26:19
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/common/canvas-tests.js:62:19
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/common/canvas-tests.js:59:15)
